Remove Dynamic Margin Calculation client
