+++ src/test/java/org/gridsuite/study/server/NetworkModificationTest.java
@@ -41,9 +41,9 @@ import org.gridsuite.study.server.repository.StudyEntity;
 import org.gridsuite.study.server.repository.StudyRepository;
 import org.gridsuite.study.server.repository.rootnetwork.RootNetworkNodeInfoRepository;
 import org.gridsuite.study.server.service.*;
-import org.gridsuite.study.server.service.client.dynamicmargincalculation.DynamicMarginCalculationClient;
 import org.gridsuite.study.server.service.client.dynamicsecurityanalysis.DynamicSecurityAnalysisClient;
 import org.gridsuite.study.server.service.client.dynamicsimulation.DynamicSimulationClient;
+import org.gridsuite.study.server.service.dynamicmargincalculation.DynamicMarginCalculationRestService;
 import org.gridsuite.study.server.service.loadflow.LoadFlowRestService;
 import org.gridsuite.study.server.service.pccmin.PccMinRestService;
 import org.gridsuite.study.server.service.securityanalysis.SecurityAnalysisRestService;
@@ -225,6 +225,9 @@ class NetworkModificationTest {
     @Autowired
     private StateEstimationRestService stateEstimationService;
 
+    @Autowired
+    DynamicMarginCalculationRestService dynamicMarginCalculationRestService;
+
     @MockitoSpyBean
     private UserAdminService userAdminService;
 
@@ -239,9 +242,6 @@ class NetworkModificationTest {
 
     @MockitoSpyBean
     DynamicSecurityAnalysisClient dynamicSecurityAnalysisClient;
-
-    @MockitoSpyBean
-    DynamicMarginCalculationClient dynamicMarginCalculationClient;
 
     @MockitoSpyBean
     private RootNetworkNodeInfoRepository rootNetworkNodeInfoRepository;
@@ -307,10 +307,10 @@ class NetworkModificationTest {
         voltageInitService.setBaseUri(baseUrl);
         stateEstimationService.setBaseUri(baseUrl);
         pccMinService.setBaseUri(baseUrl);
+        dynamicMarginCalculationRestService.setBaseUri(baseUrl);
 
         doReturn(baseUrl).when(dynamicSimulationClient).getBaseUri();
         doReturn(baseUrl).when(dynamicSecurityAnalysisClient).getBaseUri();
-        doReturn(baseUrl).when(dynamicMarginCalculationClient).getBaseUri();
 
         networkModificationService.setNetworkModificationServerBaseUri(baseUrl);
         directoryService.setDirectoryServerServerBaseUri(baseUrl);
