--- src/test/java/org/gridsuite/study/server/service/dynamicmargincalculation/DynamicMarginCalculationServiceTest.java
@@ -1,207 +1,0 @@
-/*
- * Copyright (c) 2022, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package org.gridsuite.study.server.service.dynamicmargincalculation;
-
-import org.gridsuite.study.server.ContextConfigurationWithTestChannel;
-import org.gridsuite.study.server.dto.ReportInfos;
-import org.gridsuite.study.server.dto.dynamicmargincalculation.DynamicMarginCalculationStatus;
-import org.gridsuite.study.server.service.client.dynamicmargincalculation.DynamicMarginCalculationClient;
-import org.gridsuite.study.server.utils.elasticsearch.DisableElasticsearch;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.gridsuite.study.server.StudyConstants.DYNAWO_PROVIDER;
-import static org.gridsuite.study.server.error.StudyBusinessErrorCode.COMPUTATION_RUNNING;
-import static org.gridsuite.study.server.utils.TestUtils.assertStudyException;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.*;
-
-/**
- * @author Thang PHAM <quyet-thang.pham at rte-france.com>
- */
-@SpringBootTest
-@ContextConfigurationWithTestChannel
-@DisableElasticsearch
-class DynamicMarginCalculationServiceTest {
-    private static final String VARIANT_1_ID = "variant_1";
-    private static final String PARAMETERS_JSON = "parametersJson";
-
-    // converged node
-    private static final UUID NETWORK_UUID = UUID.randomUUID();
-    private static final UUID NODE_UUID = UUID.randomUUID();
-    private static final UUID ROOTNETWORK_UUID = UUID.randomUUID();
-    private static final UUID DYNAMIC_SIMULATION_PARAMETERS_UUID = UUID.randomUUID();
-    private static final UUID DYNAMIC_SECURITY_ANALYSIS_PARAMETERS_UUID = UUID.randomUUID();
-    private static final UUID PARAMETERS_UUID = UUID.randomUUID();
-    private static final UUID DUPLICATED_PARAMETERS_UUID = UUID.randomUUID();
-    private static final UUID RESULT_UUID = UUID.randomUUID();
-    private static final UUID REPORT_UUID = UUID.randomUUID();
-
-    // running node
-    private static final UUID RESULT_UUID_RUNNING = UUID.randomUUID();
-
-    @MockitoBean
-    DynamicMarginCalculationClient dynamicMarginCalculationClient;
-    @Autowired
-    private DynamicMarginCalculationRestService dynamicMarginCalculationRestService;
-
-    @Test
-    void testGetParameters() {
-        given(dynamicMarginCalculationClient.getParameters(PARAMETERS_UUID, "userId")).willReturn(PARAMETERS_JSON);
-
-        String parametersJson = dynamicMarginCalculationRestService.getParameters(PARAMETERS_UUID, "userId");
-
-        assertThat(parametersJson).isEqualTo(PARAMETERS_JSON);
-    }
-
-    @Test
-    void testCreateParameters() {
-        given(dynamicMarginCalculationClient.createParameters(PARAMETERS_JSON)).willReturn(PARAMETERS_UUID);
-
-        UUID parametersUuid = dynamicMarginCalculationRestService.createParameters(PARAMETERS_JSON);
-
-        assertThat(parametersUuid).isEqualTo(PARAMETERS_UUID);
-    }
-
-    @Test
-    void testCreateDefaultParameters() {
-        given(dynamicMarginCalculationClient.createDefaultParameters()).willReturn(PARAMETERS_UUID);
-
-        UUID parametersUuid = dynamicMarginCalculationRestService.createDefaultParameters();
-
-        assertThat(parametersUuid).isEqualTo(PARAMETERS_UUID);
-    }
-
-    @Test
-    void testUpdateParameters() {
-        doNothing().when(dynamicMarginCalculationClient).updateParameters(PARAMETERS_UUID, PARAMETERS_JSON);
-
-        dynamicMarginCalculationRestService.updateParameters(PARAMETERS_UUID, PARAMETERS_JSON);
-
-        verify(dynamicMarginCalculationClient, times(1)).updateParameters(PARAMETERS_UUID, PARAMETERS_JSON);
-    }
-
-    @Test
-    void testDuplicateParameters() {
-        when(dynamicMarginCalculationClient.duplicateParameters(PARAMETERS_UUID)).thenReturn(DUPLICATED_PARAMETERS_UUID);
-
-        UUID newParametersUuid = dynamicMarginCalculationRestService.duplicateParameters(PARAMETERS_UUID);
-
-        assertThat(newParametersUuid).isEqualTo(DUPLICATED_PARAMETERS_UUID);
-    }
-
-    @Test
-    void testDeleteParameters() {
-        doNothing().when(dynamicMarginCalculationClient).deleteParameters(PARAMETERS_UUID);
-
-        dynamicMarginCalculationRestService.deleteParameters(PARAMETERS_UUID);
-
-        verify(dynamicMarginCalculationClient, times(1)).deleteParameters(PARAMETERS_UUID);
-    }
-
-    @Test
-    void testRunDynamicSimulation() {
-        // setup DynamicMarginCalculationClient mock
-        given(dynamicMarginCalculationClient.run(any(), eq(NETWORK_UUID), eq(VARIANT_1_ID),
-                eq(new ReportInfos(REPORT_UUID, NODE_UUID)), eq(DYNAMIC_SIMULATION_PARAMETERS_UUID), eq(DYNAMIC_SECURITY_ANALYSIS_PARAMETERS_UUID), eq(PARAMETERS_UUID), any(), eq(false)))
-                .willReturn(RESULT_UUID);
-
-        // call method to be tested
-        UUID resultUuid = dynamicMarginCalculationRestService.runDynamicMarginCalculation(NODE_UUID, ROOTNETWORK_UUID,
-                NETWORK_UUID, VARIANT_1_ID, REPORT_UUID, DYNAMIC_SIMULATION_PARAMETERS_UUID, DYNAMIC_SECURITY_ANALYSIS_PARAMETERS_UUID, PARAMETERS_UUID, "testUserId", false);
-
-        // check result
-        assertThat(resultUuid).isEqualTo(RESULT_UUID);
-    }
-
-    @Test
-    void testGetStatus() {
-        // setup DynamicMarginCalculationClient mock
-        given(dynamicMarginCalculationClient.getStatus(RESULT_UUID)).willReturn(DynamicMarginCalculationStatus.SUCCEED);
-
-        // call method to be tested
-        DynamicMarginCalculationStatus status = dynamicMarginCalculationRestService.getStatus(RESULT_UUID);
-
-        // check result
-        // status must be "SUCCEED"
-        assertThat(status).isEqualTo(DynamicMarginCalculationStatus.SUCCEED);
-    }
-
-    @Test
-    void testInvalidateStatus() {
-        List<UUID> uuids = List.of(RESULT_UUID);
-        doNothing().when(dynamicMarginCalculationClient).invalidateStatus(uuids);
-
-        dynamicMarginCalculationRestService.invalidateStatus(uuids);
-
-        verify(dynamicMarginCalculationClient, times(1)).invalidateStatus(uuids);
-    }
-
-    @Test
-    void testDeleteResult() {
-        doNothing().when(dynamicMarginCalculationClient).deleteResults(List.of(RESULT_UUID));
-
-        dynamicMarginCalculationRestService.deleteResults(List.of(RESULT_UUID));
-
-        verify(dynamicMarginCalculationClient, times(1)).deleteResults(List.of(RESULT_UUID));
-    }
-
-    @Test
-    void testDeleteResults() {
-        doNothing().when(dynamicMarginCalculationClient).deleteResults(null);
-
-        dynamicMarginCalculationRestService.deleteAllResults();
-        verify(dynamicMarginCalculationClient, times(1)).deleteResults(null);
-    }
-
-    @Test
-    void testResultCount() {
-        given(dynamicMarginCalculationClient.getResultsCount()).willReturn(10);
-
-        Integer resultsCount = dynamicMarginCalculationRestService.getResultsCount();
-
-        assertThat(resultsCount).isEqualTo(10);
-    }
-
-    @Test
-    void testAssertDynamicSecurityAnalysisNotRunning() {
-        when(dynamicMarginCalculationClient.getStatus(RESULT_UUID)).thenReturn(DynamicMarginCalculationStatus.SUCCEED);
-
-        // test not running
-        assertDoesNotThrow(() -> dynamicMarginCalculationRestService.assertDynamicMarginCalculationNotRunning(RESULT_UUID));
-
-        verify(dynamicMarginCalculationClient, times(1)).getStatus(RESULT_UUID);
-    }
-
-    @Test
-    void testAssertDynamicSecurityAnalysisRunning() {
-        // setup for running node
-        given(dynamicMarginCalculationClient.getStatus(RESULT_UUID_RUNNING)).willReturn(DynamicMarginCalculationStatus.RUNNING);
-
-        // test running
-        assertStudyException(() -> dynamicMarginCalculationRestService.assertDynamicMarginCalculationNotRunning(RESULT_UUID_RUNNING),
-            COMPUTATION_RUNNING, null);
-    }
-
-    @Test
-    void testGetProvider() {
-        given(dynamicMarginCalculationClient.getProvider(PARAMETERS_UUID)).willReturn(DYNAWO_PROVIDER);
-
-        String provider = dynamicMarginCalculationRestService.getProvider(PARAMETERS_UUID);
-
-        assertThat(provider).isEqualTo(DYNAWO_PROVIDER);
-    }
-}
