+++ src/test/java/org/gridsuite/study/server/utils/wiremock/DynamicMarginCalculationStubs.java
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.study.server.utils.wiremock;
+
+import com.github.tomakehurst.wiremock.WireMockServer;
+
+/**
+ * @author Bassel El Cheikh <bassel.el-cheikh_externe at rte-france.com>
+ */
+
+public class DynamicMarginCalculationStubs {
+    private final WireMockServer wireMock;
+
+    public DynamicMarginCalculationStubs(WireMockServer wireMock) {
+        this.wireMock = wireMock;
+    }
+}
